# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API for El Diego e-commerce application",
        "contact": {
            "email": "diegomalave@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and return access token",
                "description": "Authenticate a user and return an access token.\n\nValidates the user credentials, revokes any previous tokens,\nand issues a new one along with user information.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "diego.admin.active@e-commerce.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pasword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully authenticated."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "9|Mh1Rr1KpFbwBtUCjQkHy3qf6KqYsHjJ9SuxENCUq7e92df75"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 6
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Diego User Suspended"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "diego.user.suspended@e-commerce.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+1-240-586-8422"
                                                        },
                                                        "billing_name": {
                                                            "type": "string",
                                                            "example": "Graciela Kassulke DVM"
                                                        },
                                                        "billing_tax_id": {
                                                            "type": "string",
                                                            "example": "08934014526"
                                                        },
                                                        "billing_address_line": {
                                                            "type": "string",
                                                            "example": "521 Marilyne Mission"
                                                        },
                                                        "billing_province": {
                                                            "type": "string",
                                                            "example": "New Mexico"
                                                        },
                                                        "billing_locality": {
                                                            "type": "string",
                                                            "example": "New Antoinette"
                                                        },
                                                        "billing_zipcode": {
                                                            "type": "string",
                                                            "example": "84876"
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "suspended"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-06-23T13:06:36.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-06-23T13:06:36.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The credentials provided are invalid."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the currently authenticated user",
                "description": "Logout the authenticated user by revoking the current token.",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "You must be authenticated to access this resource."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated user profile",
                "description": "Return the currently authenticated user's information.",
                "operationId": "961ad57987905b44ae2e2f40b82e76de",
                "responses": {
                    "200": {
                        "description": "Authenticated user retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authenticated user retrieved successfully."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Diego Admin Active"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "diego.admin.active@e-commerce.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "(414) 705-9913"
                                                },
                                                "billing_name": {
                                                    "type": "string",
                                                    "example": "Willie Stoltenberg"
                                                },
                                                "billing_tax_id": {
                                                    "type": "string",
                                                    "example": "84821789858"
                                                },
                                                "billing_address_line": {
                                                    "type": "string",
                                                    "example": "672 Morar Spurs Suite 625"
                                                },
                                                "billing_province": {
                                                    "type": "string",
                                                    "example": "New Jersey"
                                                },
                                                "billing_locality": {
                                                    "type": "string",
                                                    "example": "Treutelberg"
                                                },
                                                "billing_zipcode": {
                                                    "type": "string",
                                                    "example": "41710-8957"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-23T13:06:36.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-23T13:06:36.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "You must be authenticated to access this resource."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}